refactor: extract page content components
