+++ src/views/search/SearchPageContent.tsx
@@ -1,0 +1,3 @@
+export default function SearchPageContent() {
+  return <h1>Search</h1>
+}
